add pimpl to odometry
